requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
pyodbc==5.0.1
pydantic==2.4.2
python-multipart==0.0.6 